% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_px_file.R
\name{scan_px_file}
\alias{scan_px_file}
\title{Scan a px cube file and return px data and metadata}
\usage{
scan_px_file(file_or_url, locale = NULL, encoding = "latin1")
}
\arguments{
\item{file_or_url}{url or file path to px cube
if the url gets a timeout: it is best to download
the file first and pass it as a file to the
function.}

\item{locale}{language for localizing the output (optional)
if not provided all possible translations in the px cube
are skipped}

\item{encoding}{chose encoding as either UTF-8 or latin1
default is latin1}
}
\value{
output contains of 2 parts: metadata and dataframe
}
\description{
The px file can be requested for a locale.
The data and metadata are both returned as dataframe
}
\examples{
scan_px_file(
  "https://www.pxweb.bfs.admin.ch/DownloadFile.aspx?file=px-x-1703010000_103",
  locale="fr"
)
}
