% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_px_file.R
\name{scan_px_file}
\alias{scan_px_file}
\title{Scan a px cube file and return px data and metadata}
\usage{
scan_px_file(file_or_url, locale = "default", output_dir = NULL)
}
\arguments{
\item{file_or_url}{url or file path to px cube
if the url gets a timeout: it is best to download
the file first and pass it as a file to the
function.}

\item{locale}{language for localizing the output (optional)
if not provided all possible translations in the px cube
are skipped}

\item{output_dir}{directory for writing the output files (optional)
if not provided the output will not be offered as files}
}
\value{
output contains of 2 parts: metadata and dataframe
additionally the output may be written to files
the files consist of a csv file with the data and
a json file with the metadata. In case a locale is
requested, the default metadata is provided along
with the translated metadata
}
\description{
The Px file can be requested for a locale.
The output contains the metadata as json files.
The data is returned as a dataframe but also as csv file
The output will be stored in an output directory of choice.
}
\examples{
scan_px_file("px-x-0102020203_110.px",
                       locale="en", 
                       output_dir="/tmp/")
}
